+++ IonicZlib/InfTree.cs
@@ -64,8 +64,8 @@
 using System;
 namespace Ionic.Zlib
 {
-        
-        sealed class InfTree
+
+    sealed class InfTree
         {
                 
                 private const int MANY = 1440;
@@ -274,7 +274,7 @@ namespace Ionic.Zlib
                                                         r[0] = (sbyte) j; // bits in this table
                                                         r[1] = (sbyte) l; // bits to dump before this table
                                                         j = SharedUtils.URShift(i, (w - l));
-                                                        r[2] = (int) (q - u[h - 1] - j); // offset to this table
+                                                        r[2] = q - u[h - 1] - j; // offset to this table
                                                         Array.Copy(r, 0, hp, (u[h - 1] + j) * 3, 3); // connect to last table
                                                 }
                                                 else
