+++ IonicZlib/Tree.cs
@@ -60,9 +60,6 @@
 //
 // -----------------------------------------------------------------------
 
-
-using System;
-
 namespace Ionic.Zlib
 {
     sealed class Tree
@@ -267,7 +264,7 @@ namespace Ionic.Zlib
                         continue;
                     if (tree[m * 2 + 1] != bits)
                     {
-                        s.opt_len = (int) (s.opt_len + ((long) bits - (long) tree[m * 2 + 1]) * (long) tree[m * 2]);
+                        s.opt_len = (int) (s.opt_len + ( bits - (long) tree[m * 2 + 1]) * tree[m * 2] );
                         tree[m * 2 + 1] = (short) bits;
                     }
                     n--;
